requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
ansible==6.7.0
ansible-core==2.13.13
ansible-vault==2.1.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
base58==2.1.1
bittensor==6.7.0
black==23.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.3
cytoolz==0.12.3
ddt==1.6.0
decorator==5.1.1
ecdsa==0.18.0
eth-hash==0.7.0
eth-keys==0.5.0
eth-typing==4.0.0
eth-utils==2.3.1
fastapi==0.99.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
fuzzywuzzy==0.18.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
Levenshtein==0.25.0
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy-opentensor==0.5.0
multidict==6.0.5
munch==2.5.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==1.2.1
networkx==3.2.1
numpy==1.26.4
packaging==23.2
password-strength==0.0.3.post2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
py==1.11.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.14
Pygments==2.17.2
PyNaCl==1.5.0
pytest==8.0.2
pytest-asyncio==0.23.5
python-Levenshtein==0.25.0
PyYAML==6.0.1
rapidfuzz==3.6.1
requests==2.31.0
resolvelib==0.8.1
retry==0.9.2
rich==13.7.1
scalecodec==1.2.7
shtab==1.6.5
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
substrate-interface==1.5.2
sympy==1.12
termcolor==2.4.0
toolz==0.12.1
torch==2.2.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.22.0
websocket-client==1.7.0
xxhash==3.4.1
yarl==1.9.4
python-dotenv==1.0.1
aiobotocore==2.12.3
bokeh==3.4.1